consitent renaming and refactor includes
